requirements.txt
customtkinter==5.2.2
google-generativeai==0.8.3
python-dotenv==1.0.1
Pillow==10.4.0
